requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.1.0
celery==5.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
boto3==1.35.0
pgvector==0.3.0
numpy==1.26.0
anthropic==0.39.0
langgraph==0.2.0
langchain-anthropic==0.2.0
python-dotenv==1.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
